% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chgamma}
\alias{chgamma}
\title{Gamma Cumulative hazard}
\usage{
chgamma(t, shape, scale)
}
\arguments{
\item{t:}{positive parameter}

\item{shape:}{shape parameter}

\item{scale:}{scale parameter}

\item{log:}{return log harzard (TRUE or FALSE)}
}
\value{
Gamma cumulative hazard function
}
\description{
Gamma Cumulative hazard
}
