% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chlnorm}
\alias{chlnorm}
\title{Lognormal Cumulative hazard}
\usage{
chlnorm(t, mu, sigma)
}
\arguments{
\item{t:}{positive parameter}

\item{mu:}{log-location parameter}

\item{sigma:}{scale parameter}
}
\value{
lognormal cumulative hazard function
}
\description{
Lognormal Cumulative hazard
}
