% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chpgw}
\alias{chpgw}
\title{PGW Cumulative Hazard Function}
\usage{
chpgw(t, sigma, nu, gamma)
}
\arguments{
\item{t:}{positive parameter}

\item{sigma:}{scale parameter}

\item{nu:}{shape parameter}

\item{gamma:}{shape parameter}
}
\value{
PGW cumulative hazard function
}
\description{
PGW Cumulative Hazard Function
}
