% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chw}
\alias{chw}
\title{Weibull Cumulative hazard}
\usage{
chw(t, scale, shape)
}
\arguments{
\item{t:}{positive value}

\item{scale:}{scale parameter}

\item{shape:}{shape parameter}
}
\value{
Weibull cumulative hazard function.
}
\description{
Weibull Cumulative hazard
}
