% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{CI.NSfit_IFPGWGH}
\alias{CI.NSfit_IFPGWGH}
\title{Confidence intervals for Net survival at time t0: IFPGWGH model}
\usage{
CI.NSfit_IFPGWGH(
  t0,
  NMC,
  MLE,
  level,
  times,
  rates,
  status,
  des_t,
  des,
  subgroup
)
}
\arguments{
\item{t0}{: time at what the confidence interval will be calculated}

\item{NMC}{: number of Monte Carlo iterations}

\item{level}{: confidence level}

\item{MLE:}{Maximum likelihood estimate}

\item{times:}{survival times}

\item{status:}{vital status}

\item{rates:}{population hazard values}

\item{des:}{design matrix for hazard-level effects}

\item{des_t:}{design matrix for time-level effects}

\item{subgroup:}{indices of subgroup of interest}
}
\value{
matrix containing ("lower","estimate","upper")
}
\description{
Confidence intervals for Net survival at time t0: IFPGWGH model
}
