% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{Conf.Int}
\alias{Conf.Int}
\title{Function to calculate the normal confidence intervals on the log scale for positive par}
\usage{
Conf.Int(FUN, MLE, level = 0.95)
}
\arguments{
\item{FUN}{: minus log-likelihood function to be used to calculate the confidence intervals}

\item{MLE}{: maximum likelihood estimator of the parameters of interest}

\item{level}{: confidence level}
}
\value{
matrix with colums ("Lower","Upper","Transf MLE", "Std. Error")
}
\description{
Function to calculate the normal confidence intervals on the log scale for positive par
}
