% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{Conf.IntPGWGH}
\alias{Conf.IntPGWGH}
\title{Function to calculate std errors and inclusion of the true parameter values}
\usage{
Conf.IntPGWGH(
  trans.mle,
  level = 0.95,
  times,
  status,
  rates,
  des = NULL,
  des_t = NULL
)
}
\arguments{
\item{trans.mle}{: mle with log scale for positive parameters}

\item{level:}{confidence level}

\item{times:}{survival times}

\item{status:}{vital status}

\item{rates:}{population hazard values}

\item{des:}{design matrix for hazard-level effects}

\item{des_t:}{design matrix for time-level effects}
}
\value{
confidence intervals for PGWGH model
}
\description{
Function to calculate std errors and inclusion of the true parameter values
}
