% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{hgamma}
\alias{hgamma}
\title{Gamma Hazard}
\usage{
hgamma(t, shape, scale, log = FALSE)
}
\arguments{
\item{t:}{positive parameter}

\item{shape:}{shape parameter}

\item{scale:}{scale parameter}

\item{log:}{return log harzard (TRUE or FALSE)}
}
\value{
Gamma hazard function
}
\description{
Gamma Hazard
}
