% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{hlnorm}
\alias{hlnorm}
\title{Lognormal Hazard}
\usage{
hlnorm(t, mu, sigma, log = FALSE)
}
\arguments{
\item{t:}{positive parameter}

\item{mu:}{log-location parameter}

\item{sigma:}{scale parameter}

\item{log:}{return log harzard (TRUE or FALSE)}
}
\value{
lognormal hazard function
}
\description{
Lognormal Hazard
}
