% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{hpgw}
\alias{hpgw}
\title{PGW Hazard Function}
\usage{
hpgw(t, sigma, nu, gamma, log = FALSE)
}
\arguments{
\item{t:}{positive parameter}

\item{sigma:}{scale parameter}

\item{nu:}{shape parameter}

\item{gamma:}{shape parameter}

\item{log:}{return log hazard (TRUE or FALSE)}
}
\value{
PGW hazard function
}
\description{
PGW Hazard Function
}
