% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{hw}
\alias{hw}
\title{Weibull Hazard}
\usage{
hw(t, scale, shape)
}
\arguments{
\item{t:}{positive value}

\item{scale:}{scale parameter}

\item{shape:}{shape parameter}
}
\value{
Weibull hazard function.
}
\description{
Weibull Hazard
}
