% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{IFLNMLE}
\alias{IFLNMLE}
\title{Individual Fraitly Excess Hazard Models: LN Baseline hazard
IFLNMLE function. It returns the output from optim or nlminb}
\usage{
IFLNMLE(
  init,
  hstr = "IFLN",
  times,
  status,
  rates,
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
(log-location, log(scale), alpha, beta, log(b)), 
where alpha is only required for LNGH}

\item{hstr}{: hazard structure, LN without covariates ("IFLN"),
LN with AFT model (IFLNAFT)
LN with PH model (IFLNPH)
LN with AH model (IFLNAH)
LN with GH model (IFLNGH)}

\item{times}{: times to event}

\item{status}{: status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{rates}{: expected population mortality hazard rates}

\item{method}{: optimisation method to be used in optim}

\item{maxit}{: maximum number of iterations in optim}
}
\value{
list containing the output from the optimsation (OPT) and the log-likelihood (loglik)
}
\description{
Individual Fraitly Excess Hazard Models: LN Baseline hazard
IFLNMLE function. It returns the output from optim or nlminb
}
