% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{IFNS}
\alias{IFNS}
\title{Function to evaluate the net survival based on specific values of the parameters and covariates}
\usage{
IFNS(
  t,
  sigma,
  nu,
  gamma,
  alpha = NULL,
  beta = NULL,
  scale.frail,
  des = NULL,
  des_t = NULL,
  hstr = "IFPGW"
)
}
\arguments{
\item{t:}{positive value}

\item{sigma:}{scale parameter}

\item{nu:}{shape parameter}

\item{gamma:}{shape parameter}

\item{alpha:}{regression parameters for time-level effects}

\item{beta:}{regression coefficients for hazard-level effects}

\item{scale.frail:}{scale parameter of the frailty distribution}

\item{des:}{design matrix for hazard-level effects}

\item{des_t:}{design matrix for time-level effects}
}
\value{
Net survival for IFPGW model: IFPGW, IFPGWPH, IFPGWAH, IFPGWAFT, IFPGWGH
}
\description{
Function to evaluate the net survival based on specific values of the parameters and covariates
}
