% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{IFPGWMLE}
\alias{IFPGWMLE}
\title{Individual Fraitly Excess Hazard Models: PGW Baseline hazard
IFPGWMLE function. It returns the output from optim or nlminb}
\usage{
IFPGWMLE(
  init,
  hstr = "IFPGW",
  times,
  status,
  rates,
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
(log(scale), log(shape1), log(shape2), alpha, beta, log(b)), 
where alpha is only required for PGWGH}

\item{hstr}{: hazard structure, PGW without covariates ("IFPGW"),
PGW with AFT model (IFPGWAFT)
PGW with PH model (IFPGWPH)
PGW with AH model (IFPGWAH)
PGW with GH model (IFPGWGH)}

\item{times}{: times to event}

\item{status}{: status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{rates}{: expected population mortality hazard rates}

\item{method}{: optimisation method to be used in optim}

\item{maxit}{: maximum number of iterations in optim}
}
\value{
list containing the output from the optimsation (OPT) and the log-likelihood (loglik)
}
\description{
Individual Fraitly Excess Hazard Models: PGW Baseline hazard
IFPGWMLE function. It returns the output from optim or nlminb
}
