% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{LNMLE}
\alias{LNMLE}
\title{Excess Hazard Models: LogNormal baseline with GH structure
LNMLE function. It returns the output from optim or nlminb}
\usage{
LNMLE(
  init,
  times,
  status,
  rates,
  hstr = "LN",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
(log-location, log(scale), alpha, beta), 
where alpha is only required for LNGH}

\item{times}{: times to event}

\item{status}{: status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{rates}{: expected population mortality hazard rates}

\item{hstr}{: hazard structure, LN without covariates ("LN"),
LN with AFT model (LNAFT)
LN with PH model (LNPH)
LN with AH model (LNAH)
LN with GH model (LNGH)}

\item{des}{: design matrix for proportional hazard effects}

\item{des_t}{: design matrix for time-dependent effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim}

\item{maxit}{: maximum number of iterations in optim}
}
\value{
list containing the output from the optimsation (OPT) and the log-likelihood (loglik)
}
\description{
Excess Hazard Models: LogNormal baseline with GH structure
LNMLE function. It returns the output from optim or nlminb
}
