% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{NSfit_IFLNGH}
\alias{NSfit_IFLNGH}
\title{Net survival function for IFLNGH model}
\usage{
NSfit_IFLNGH(t, par, des_t, des)
}
\arguments{
\item{t:}{positive value}

\item{par:}{model parameters}

\item{des:}{design matrix for hazard-level effects}

\item{des_t:}{design matrix for time-level effects}
}
\value{
net survival for IFLNGH model
}
\description{
Net survival function for IFLNGH model
}
