% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{NSfit_IFPGWPH}
\alias{NSfit_IFPGWPH}
\title{Net survival function for IFPGWPH model}
\usage{
NSfit_IFPGWPH(t, par, des)
}
\arguments{
\item{t:}{positive value}

\item{par:}{model parameters}

\item{des:}{design matrix for hazard-level effects}
}
\value{
net survival for IFPGWPH model
}
\description{
Net survival function for IFPGWPH model
}
