% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{NSfit_LNGH}
\alias{NSfit_LNGH}
\title{Net survival function for LNGH model}
\usage{
NSfit_LNGH(t, par, des_t, des)
}
\arguments{
\item{t:}{positive value}

\item{par:}{model parameters}

\item{des:}{design matrix for hazard-level effects}

\item{des_t:}{design matrix for time-level effects}
}
\value{
net survival for LNGH model
}
\description{
Net survival function for LNGH model
}
