% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{NSfit_PGWPH}
\alias{NSfit_PGWPH}
\title{Net survival function for PGWPH model}
\usage{
NSfit_PGWPH(t, par, des)
}
\arguments{
\item{t:}{positive value}

\item{par:}{model parameters}

\item{des:}{design matrix for hazard-level effects}
}
\value{
net survival for PGWPH model
}
\description{
Net survival function for PGWPH model
}
