% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{PGWMLE}
\alias{PGWMLE}
\title{Excess Hazard Models: PGW baseline with GH structure
PGWMLE function. It returns the output from optim or nlminb}
\usage{
PGWMLE(
  init,
  times,
  status,
  rates,
  hstr = "PGW",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
(log(scale), log(shape1), log(shape2), alpha, beta), 
           where alpha is only required for PGWGH}

\item{times}{: times to event}

\item{status}{: status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{rates}{: expected population mortality hazard rates}

\item{hstr}{: hazard structure, PGW without covariates ("PGW"),
PGW with AFT model (PGWAFT)
PGW with PH model (PGWPH)
PGW with AH model (PGWAH)
PGW with GH model (PGWGH)}

\item{des}{: design matrix for proportional hazard effects}

\item{des_t}{: design matrix for time-dependent effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim}

\item{maxit}{: maximum number of iterations in optim}
}
\value{
list containing the output from the optimsation (OPT) and the log-likelihood (loglik)
}
\description{
Excess Hazard Models: PGW baseline with GH structure
PGWMLE function. It returns the output from optim or nlminb
}
