% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{qpgw}
\alias{qpgw}
\title{Quantile Function}
\usage{
qpgw(p, sigma, nu, gamma)
}
\arguments{
\item{p:}{probability value}

\item{sigma:}{scale parameter}

\item{nu:}{shape parameter}

\item{gamma:}{shape parameter}
}
\value{
vector of quantiles
}
\description{
Quantile Function
}
