% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{rpgw}
\alias{rpgw}
\title{Random Number Generation Function PGW distribution}
\usage{
rpgw(n, sigma, nu, gamma)
}
\arguments{
\item{n:}{sample size}

\item{sigma:}{scale parameter}

\item{nu:}{shape parameter}

\item{gamma:}{shape parameter}
}
\value{
vector of simulated values
}
\description{
Random Number Generation Function PGW distribution
}
