% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{simLNGH}
\alias{simLNGH}
\title{Function to simulate times to event from a model with LN baseline hazard
and GH structure}
\usage{
simLNGH(seed, n, X, S, parLN, alpha, beta, frailty = "One", scale.frail = NULL)
}
\arguments{
\item{seed:}{simulation seed}

\item{n:}{number of simulations}

\item{des_t:}{design matrix for time-level effects}

\item{des:}{design matrix for hazard-level effects}

\item{parLN:}{parameters of the lognormal baseline hazard}

\item{alpha:}{regression coefficients for time-level effects}

\item{beta:}{regression coefficients for hazard-level effects}

\item{frailty:}{"One" indicates frailties are one (no frailty), "Gamma" indicates gamma-distributed frailties}

\item{scale.frail:}{scale parameter for the frailty distribution}
}
\value{
returns the vector "lifestimes" containing the simulated survival times.
}
\description{
Function to simulate times to event from a model with LN baseline hazard
and GH structure
}
