% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{simPGWGH}
\alias{simPGWGH}
\title{Function to simulate times to event from a model with PGW baseline hazard
and GH structure}
\usage{
simPGWGH(
  seed,
  n,
  des_t,
  des,
  parPGW,
  alpha,
  beta,
  frailty = "One",
  scale.frail = NULL
)
}
\arguments{
\item{seed:}{simulation seed}

\item{n:}{number of simulations}

\item{des_t:}{design matrix for time-level effects}

\item{des:}{design matrix for hazard-level effects}

\item{parPGW:}{parameters of the PGW baseline hazard}

\item{alpha:}{regression coefficients for time-level effects}

\item{beta:}{regression coefficients for hazard-level effects}

\item{frailty:}{"One" indicates frailties are one (no frailty), "Gamma" indicates gamma-distributed frailties}

\item{scale.frail:}{scale parameter for the frailty distribution}
}
\value{
returns the vector "lifestimes" containing the simulated survival times.
}
\description{
Function to simulate times to event from a model with PGW baseline hazard
and GH structure
}
