% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{spgw}
\alias{spgw}
\title{PGW Survival Function}
\usage{
spgw(t, sigma, nu, gamma, log.p = FALSE)
}
\arguments{
\item{t:}{positive parameter}

\item{sigma:}{scale parameter}

\item{nu:}{shape parameter}

\item{gamma:}{shape parameter}

\item{log.p:}{return log survival (TRUE or FALSE)}
}
\value{
PGW survival function
}
\description{
PGW Survival Function
}
