% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{WMLE}
\alias{WMLE}
\title{Excess Hazard Models: Weibull baseline
WMLE function. It returns the output from optim or nlminb}
\usage{
WMLE(
  init,
  hstr = "W",
  times,
  status,
  rates,
  des = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step   (log(scale), log(shape), beta),}

\item{hstr}{: hazard structure, Weibull without covariates ("W")
Weibull with AFT model (WAFT)
Weibull with PH model (WPH)
Weibull with AH model (WAH)}

\item{times}{: times to event}

\item{status}{: status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{rates}{: expected population mortality hazard rates}

\item{des}{: design matrix}

\item{method}{: "nlminb" or optimisation method to be used in optim}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
WMLE returns an object containing the output of optim or nlminb
}
\description{
Excess Hazard Models: Weibull baseline
WMLE function. It returns the output from optim or nlminb
}
